% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{etl_extract.etl_etllahman}
\alias{etl_extract.etl_etllahman}
\alias{etl_transform.etl_etllahman}
\title{etl_extract.etl_etllahman}
\usage{
\method{etl_extract}{etl_etllahman}(obj, ...)

\method{etl_transform}{etl_etllahman}(obj, ...)
}
\description{
etl_extract.etl_etllahman
}
