% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl.R
\name{etl_extract.etl_etllahman}
\alias{etl_extract.etl_etllahman}
\alias{etl_transform.etl_etllahman}
\alias{etl_load.etl_etllahman}
\title{etl_extract.etl_etllahman
ETL methods for accessing the Lahman Database}
\source{
\url{http://www.seanlahman.com/baseball-archive/statistics/}
}
\usage{
\method{etl_extract}{etl_etllahman}(obj, ...)

\method{etl_transform}{etl_etllahman}(obj, ...)

\method{etl_load}{etl_etllahman}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
etl_extract.etl_etllahman
ETL methods for accessing the Lahman Database
}
\seealso{
\code{\link[etl]{etl_extract}}
}
